// src/app/models/vehicle-drivers/driver.model.ts
export interface Driver {
  _id?: string;             // Optional, generated by backend
  name: string;            // Driver’s full name
  status: 'Available' | 'On Trek';  // Status of the driver
  returnDate?: string | null; // Return date if on trek, otherwise null
  updatedAt?: string | Date;
}